2021.12.28 추가 작업

+++ app/src/main/java/com/example/takealook/Best/FragmentBestToday.kt
@@ -69,7 +69,9 @@ class FragmentBestToday : Fragment() {
         ).allowMainThreadQueries()
             .build()
 
-        db.bestDao().deleteWeek(day)
+        if(db.bestDao().getAll() != null){
+            db.bestDao().deleteWeek(day)
+        }
 
         return root
     }
